# Conflicts:
#   ItsGrey/app/build.gradle.kts
#   ItsGrey/app/src/main/java/com/tntt/itsgrey/MainActivity.kt
#   ItsGrey/app/src/main/java/com/tntt/itsgrey/navigation/IgNavHost.kt

+++ ItsGrey/data/book/src/main/java/com/tntt/book/datasource/RemoteBookDataSourceImpl.kt
@@ -61,7 +61,7 @@ class RemoteBookDataSourceImpl @Inject constructor(
                 .whereEqualTo("userId", userId)
                 .whereEqualTo("bookType", bookType)
                 .orderBy("title", Query.Direction.ASCENDING)
-                .limit((startIndex + 1L))
+                .limit((startIndex + 1).toLong())
                 .get()
                 .await()
